Cube demo works now, working on rsxtest.elf

+++ src/kernel/Memory.h
@@ -20,7 +20,7 @@ private:
 
     struct Block
     {
-        uint64_t addr;
+        uint32_t addr;
         uint64_t end;
     };
 
@@ -29,7 +29,7 @@ public:
     MemoryBlock(uint64_t start, uint64_t end, MemoryManager* manager, bool map = true);
 
     uint64_t Alloc(uint64_t size);
-    void MarkUsed(uint64_t addr, uint64_t size);
+    void MarkUsed(uint32_t addr, uint64_t size);
 
     uint64_t GetStart() const {return begin;}
     uint64_t GetSize() const {return len;}
@@ -48,22 +48,22 @@ public:
     MemoryManager();
     ~MemoryManager();
 
-    void Write8(uint64_t addr, uint8_t data, bool slow = false);
-    void Write16(uint64_t addr, uint16_t data, bool slow = false);
-    void Write32(uint64_t addr, uint32_t data, bool slow = false);
-    void Write64(uint64_t addr, uint64_t data, bool slow = false);
+    void Write8(uint32_t addr, uint8_t data, bool slow = false);
+    void Write16(uint32_t addr, uint16_t data, bool slow = false);
+    void Write32(uint32_t addr, uint32_t data, bool slow = false);
+    void Write64(uint32_t addr, uint64_t data, bool slow = false);
 
-    uint8_t Read8(uint64_t addr, bool slow = false);
-    uint16_t Read16(uint64_t addr, bool slow = false);
-    uint32_t Read32(uint64_t addr, bool slow = false);
-    uint64_t Read64(uint64_t addr, bool slow = false);
+    uint8_t Read8(uint32_t addr, bool slow = false);
+    uint16_t Read16(uint32_t addr, bool slow = false);
+    uint32_t Read32(uint32_t addr, bool slow = false);
+    uint64_t Read64(uint32_t addr, bool slow = false);
 
     void MarkMemoryRegion(uint64_t start, uint64_t end, int flags); // 1 = R, 2 = W
     uint8_t* GetRawPtr(uint64_t offset);
 
     void MapMemory(uint64_t start, uint64_t end, uint8_t* ptr);
     
-    void SetRSXControlReg(uint64_t addr);
+    void SetRSXControlReg(uint32_t addr);
 
     void DumpRam();
